+++ temba/orgs/tests.py
@@ -22,7 +22,7 @@ from temba.airtime.models import AirtimeTransfer
 from temba.api.models import APIToken, Resthook, WebHookEvent
 from temba.archives.models import Archive
 from temba.campaigns.models import Campaign, CampaignEvent, EventFire
-from temba.channels.models import Alert, Channel, SyncEvent
+from temba.channels.models import Alert, Channel, ChannelLog, SyncEvent
 from temba.classifiers.models import Classifier
 from temba.classifiers.types.wit import WitType
 from temba.contacts.models import (
@@ -35,7 +35,7 @@ from temba.contacts.models import (
     ContactURN,
     ExportContactsTask,
 )
-from temba.flows.models import ExportFlowResultsTask, Flow, FlowLabel, FlowRun, FlowSession, FlowStart
+from temba.flows.models import ExportFlowResultsTask, Flow, FlowLabel, FlowRun, FlowSession, FlowStart, FlowStartCount
 from temba.globals.models import Global
 from temba.locations.models import AdminBoundary
 from temba.msgs.models import ExportMessagesTask, Label, Msg
@@ -79,21 +79,21 @@ class OrgContextProcessorTest(TembaTest):
         self.assertTrue(perms["contacts"]["contact_update"])
         self.assertTrue(perms["orgs"]["org_country"])
         self.assertTrue(perms["orgs"]["org_manage_accounts"])
-        self.assertTrue(perms["orgs"]["org_delete"])
+        self.assertTrue(perms["orgs"]["org_delete_child"])
 
         perms = RolePermsWrapper(OrgRole.EDITOR)
 
         self.assertTrue(perms["msgs"]["msg_list"])
         self.assertTrue(perms["contacts"]["contact_update"])
         self.assertFalse(perms["orgs"]["org_manage_accounts"])
-        self.assertFalse(perms["orgs"]["org_delete"])
+        self.assertFalse(perms["orgs"]["org_delete_child"])
 
         perms = RolePermsWrapper(OrgRole.VIEWER)
 
         self.assertTrue(perms["msgs"]["msg_list"])
         self.assertFalse(perms["contacts"]["contact_update"])
         self.assertFalse(perms["orgs"]["org_manage_accounts"])
-        self.assertFalse(perms["orgs"]["org_delete"])
+        self.assertFalse(perms["orgs"]["org_delete_child"])
 
         self.assertFalse(perms["msgs"]["foo"])  # no blow up if perm doesn't exist
         self.assertFalse(perms["chickens"]["foo"])  # or app doesn't exist
@@ -2152,7 +2152,7 @@ class OrgDeleteTest(TembaTest):
         flows = self._create_flow_content(org, user, channels, contacts, groups, add)
         labels = self._create_message_content(org, user, channels, contacts, groups, add)
         self._create_campaign_content(org, user, fields, groups, flows, contacts, add)
-        self._create_ticket_content(org, user, contacts, add)
+        self._create_ticket_content(org, user, contacts, flows, add)
         self._create_export_content(org, user, flows, groups, fields, labels, add)
         self._create_archive_content(org, add)
 
@@ -2169,31 +2169,53 @@ class OrgDeleteTest(TembaTest):
     def _create_channel_content(self, org, add) -> tuple:
         channel1 = add(self.create_channel("TG", "Telegram", "+250785551212", org=org))
         channel2 = add(self.create_channel("A", "Android", "+1234567890", org=org))
-        SyncEvent.create(
-            channel2,
-            dict(pending=[], retry=[], power_source="P", power_status="full", power_level="100", network_type="W"),
-            [],
+        channel3 = add(self.create_channel("T", "Twilio", "+1098765432", parent=channel2, org=org))
+        add(
+            SyncEvent.create(
+                channel2,
+                dict(pending=[], retry=[], power_source="P", power_status="full", power_level="100", network_type="W"),
+                [],
+            )
         )
-        Alert.objects.create(
-            channel=channel2, alert_type=Alert.TYPE_POWER, created_by=self.admin, modified_by=self.admin
+        add(
+            Alert.objects.create(
+                channel=channel2, alert_type=Alert.TYPE_POWER, created_by=self.admin, modified_by=self.admin
+            )
+        )
+        add(ChannelLog.objects.create(channel=channel1, log_type=ChannelLog.LOG_TYPE_MSG_SEND))
+        add(
+            HTTPLog.objects.create(
+                org=org, channel=channel2, log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED, request_time=10, is_error=False
+            )
         )
 
-        return (channel1, channel2)
+        return (channel1, channel2, channel3)
 
     def _create_flow_content(self, org, user, channels, contacts, groups, add) -> tuple:
         flow1 = add(self.create_flow("Registration", org=org))
         flow2 = add(self.create_flow("Goodbye", org=org))
 
-        add(FlowStart.objects.create(org=org, flow=flow1))
+        start1 = add(FlowStart.objects.create(org=org, flow=flow1))
+        add(FlowStartCount.objects.create(start=start1, count=1))
+
         add(
             Trigger.create(
                 org,
                 user,
                 flow=flow1,
                 trigger_type=Trigger.TYPE_KEYWORD,
-                channel=channels[0],
                 keyword="color",
                 match_type=Trigger.MATCH_FIRST_WORD,
+                groups=groups,
+            )
+        )
+        add(
+            Trigger.create(
+                org,
+                user,
+                flow=flow1,
+                trigger_type=Trigger.TYPE_NEW_CONVERSATION,
+                channel=channels[0],
                 groups=groups,
             )
         )
@@ -2220,6 +2242,8 @@ class OrgDeleteTest(TembaTest):
                 exited_on=timezone.now(),
             )
         )
+        contacts[0].current_flow = flow1
+        contacts[0].save(update_fields=("current_flow",))
 
         flow_label1 = add(FlowLabel.create(org, user, "Cool Flows"))
         flow_label2 = add(FlowLabel.create(org, user, "Crazy Flows"))
@@ -2321,13 +2345,15 @@ class OrgDeleteTest(TembaTest):
             )
         )
         add(EventFire.objects.create(event=event1, contact=contacts[0], scheduled=timezone.now()))
+        start1 = add(FlowStart.objects.create(org=org, flow=flows[0], campaign_event=event1))
+        add(FlowStartCount.objects.create(start=start1, count=1))
 
-    def _create_ticket_content(self, org, user, contacts, add):
+    def _create_ticket_content(self, org, user, contacts, flows, add):
         ticket1 = add(self.create_ticket(org.ticketers.get(), contacts[0], "Help"))
         ticket1.events.create(org=org, contact=contacts[0], event_type="N", note="spam", created_by=user)
 
         ticketer = add(Ticketer.create(org, user, MailgunType.slug, "Email (bob)", {}))
-        add(self.create_ticket(ticketer, contacts[0], "Help"))
+        add(self.create_ticket(ticketer, contacts[0], "Help", opened_in=flows[0]))
 
     def _create_export_content(self, org, user, flows, groups, fields, labels, add):
         results = add(
@@ -2366,7 +2392,7 @@ class OrgDeleteTest(TembaTest):
             body, md5, size = jsonlgz_encode([{"id": 1}])
             archive = Archive.objects.create(
                 org=org,
-                url=f"http://{settings.STORAGE_BUCKETS['archives']}.aws.com/{file}",
+                url=f"http://{settings.ARCHIVE_BUCKET}.aws.com/{file}",
                 start_date=timezone.now(),
                 build_time=100,
                 archive_type=Archive.TYPE_MSG,
@@ -2375,14 +2401,14 @@ class OrgDeleteTest(TembaTest):
                 size=size,
                 hash=md5,
             )
-            self.mock_s3.put_object(settings.STORAGE_BUCKETS["archives"], file, body)
+            self.mock_s3.put_object(settings.ARCHIVE_BUCKET, file, body)
             return archive
 
         daily = add(create_archive(org, Archive.PERIOD_DAILY))
         add(create_archive(org, Archive.PERIOD_MONTHLY, daily))
 
         # extra S3 file in child archive dir
-        self.mock_s3.put_object(settings.STORAGE_BUCKETS["archives"], f"{org.id}/extra_file.json", io.StringIO("[]"))
+        self.mock_s3.put_object(settings.ARCHIVE_BUCKET, f"{org.id}/extra_file.json", io.StringIO("[]"))
 
     def _exists(self, obj) -> bool:
         return obj._meta.model.objects.filter(id=obj.id).exists()
@@ -2646,16 +2672,14 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertEqual(1, len(response.context["children"]))
 
         # we should have options to flag and suspend
-        self.assertContentMenu(
-            read_url, self.customer_support, ["Edit", "Flag", "Suspend", "Verify", "-", "Delete", "-", "Service"]
-        )
+        self.assertContentMenu(read_url, self.customer_support, ["Edit", "Flag", "Suspend", "Verify", "-", "Service"])
 
         # flag and content menu option should be inverted
         self.org.flag()
         self.org.suspend()
 
         self.assertContentMenu(
-            read_url, self.customer_support, ["Edit", "Unflag", "Unsuspend", "Verify", "-", "Delete", "-", "Service"]
+            read_url, self.customer_support, ["Edit", "Unflag", "Unsuspend", "Verify", "-", "Service"]
         )
 
         # no menu for inactive orgs
@@ -3438,43 +3462,25 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertEqual("Y", self.org.date_format)
         self.assertEqual("es", self.org.language)
 
-    def test_delete(self):
+    def test_delete_child(self):
         self.org.features = [Org.FEATURE_CHILD_ORGS]
         self.org.save(update_fields=("features",))
 
-        workspace = self.org.create_new(self.admin, "Child Workspace", self.org.timezone, as_child=True)
-        delete_workspace = reverse("orgs.org_delete", args=[workspace.id])
+        child = self.org.create_new(self.admin, "Child Workspace", self.org.timezone, as_child=True)
+        delete_url = reverse("orgs.org_delete_child", args=[child.id])
 
-        # choose the parent org, try to delete the workspace
-        self.assertDeleteFetch(delete_workspace)
+        self.assertDeleteFetch(delete_url)
 
         # schedule for deletion
-        response = self.client.get(delete_workspace)
+        response = self.client.get(delete_url)
         self.assertContains(response, "You are about to delete the workspace <b>Child Workspace</b>")
 
         # go through with it, redirects to main workspace page
-        response = self.client.post(delete_workspace)
+        response = self.client.post(delete_url)
         self.assertEqual(reverse("orgs.org_sub_orgs"), response["Temba-Success"])
 
-        workspace.refresh_from_db()
-        self.assertFalse(workspace.is_active)
-
-        # can't delete primary workspace
-        primary_delete = reverse("orgs.org_delete", args=[self.org.id])
-        response = self.client.get(primary_delete)
-        self.assertRedirect(response, "/users/login/")
-
-        response = self.client.post(primary_delete)
-        self.assertRedirect(response, "/users/login/")
-
-        self.login(self.customer_support)
-        primary_delete = reverse("orgs.org_delete", args=[self.org.id])
-        response = self.client.get(primary_delete)
-        self.assertContains(response, "You are about to delete the workspace <b>Nyaruka</b>")
-
-        response = self.client.post(primary_delete)
-        self.org.refresh_from_db()
-        self.assertFalse(self.org.is_active)
+        child.refresh_from_db()
+        self.assertFalse(child.is_active)
 
     def test_administration(self):
         self.setUpLocations()