+++ temba/orgs/views.py
@@ -35,6 +35,7 @@ from django.contrib.auth.views import LoginView as AuthLoginView
 from django.core.exceptions import ValidationError
 from django.core.validators import validate_email
 from django.db import IntegrityError
+from django.db.models.functions import Lower
 from django.forms import ModelChoiceField
 from django.http import Http404, HttpResponse, HttpResponseRedirect, JsonResponse
 from django.shortcuts import resolve_url
@@ -47,8 +48,6 @@ from django.views.decorators.csrf import csrf_exempt
 
 from temba.api.models import APIToken, Resthook
 from temba.campaigns.models import Campaign
-from temba.channels.models import Channel
-from temba.classifiers.models import Classifier
 from temba.flows.models import Flow
 from temba.formax import FormaxMixin
 from temba.utils import analytics, get_anonymous_user, json, languages
@@ -1134,7 +1133,7 @@ class OrgCRUDL(SmartCRUDL):
         "grant",
         "create_login",
         "choose",
-        "delete",
+        "delete_child",
         "manage_accounts",
         "manage_accounts_sub_org",
         "manage",
@@ -1244,11 +1243,11 @@ class OrgCRUDL(SmartCRUDL):
                     from temba.channels.views import get_channel_read_url
 
                     items = []
-                    channels = Channel.objects.filter(org=self.org, is_active=True, parent=None).order_by("-role")
+                    channels = self.org.channels.filter(is_active=True, parent=None).order_by(Lower("name"))
                     for channel in channels:
                         items.append(
                             self.create_menu_item(
-                                menu_id=f"{channel.uuid}",
+                                menu_id=str(channel.uuid),
                                 name=channel.name,
                                 href=get_channel_read_url(channel),
                                 icon=channel.type.get_icon(),
@@ -1260,7 +1259,7 @@ class OrgCRUDL(SmartCRUDL):
 
                 if self.has_org_perm("classifiers.classifier_read"):
                     items = []
-                    classifiers = Classifier.objects.filter(org=self.org, is_active=True).order_by("-created_on")
+                    classifiers = self.org.classifiers.filter(is_active=True).order_by(Lower("name"))
                     for classifier in classifiers:
                         items.append(
                             self.create_menu_item(
@@ -1723,15 +1722,6 @@ class OrgCRUDL(SmartCRUDL):
                 menu.add_url_post(_("Verify"), f"{reverse('orgs.org_update', args=[obj.id])}?action=verify")
 
             menu.new_group()
-            menu.add_modax(
-                _("Delete"),
-                "delete-org",
-                reverse("orgs.org_delete", args=[obj.id]),
-                title=_("Delete Workspace"),
-                disabled=True,
-            )
-
-            menu.new_group()
             menu.add_url_post(
                 _("Service"),
                 f'{reverse("orgs.org_service")}?other_org={obj.id}&next={reverse("msgs.msg_inbox", args=[])}',
@@ -1905,23 +1895,11 @@ class OrgCRUDL(SmartCRUDL):
             obj.limits = cleaned_data["limits"]
             return obj
 
-    class Delete(SpaMixin, OrgObjPermsMixin, ModalMixin, SmartDeleteView):
-        cancel_url = "id@orgs.org_update"
+    class DeleteChild(SpaMixin, OrgObjPermsMixin, ModalMixin, SmartDeleteView):
+        cancel_url = "@orgs.org_sub_orgs"
         success_url = "@orgs.org_sub_orgs"
         fields = ("id",)
         submit_button_name = _("Delete")
-
-        # we don't want to reroute delete requests
-        def pre_process(self, request, *args, **kwargs):
-            return
-
-        def has_org_perm(self, codename):
-            # users can't delete the primary org
-            org = self.get_object()
-            if not org.is_child:
-                return False
-
-            return super().has_org_perm(codename)
 
         def get_object_org(self):
             # child orgs work in the context of their parent
@@ -1934,6 +1912,8 @@ class OrgCRUDL(SmartCRUDL):
             return context
 
         def post(self, request, *args, **kwargs):
+            assert self.get_object().is_child, "can only delete child orgs"
+
             self.object = self.get_object()
             self.object.release(request.user)
             return self.render_modal_response()