+++ temba/utils/models/base.py
@@ -25,6 +25,32 @@ def patch_queryset_count(qs, function):
     qs.count = types.MethodType(lambda s: function(), qs)
 
 
+def delete_in_batches(qs, *, batch_size: int = 1000, pk: str = "id", pre_delete=None, post_delete=None) -> int:
+    """
+    Deletes objects from the given queryset in batches returning the number deleted. Callback functions can be provided
+    as `pre_delete` and `post_delete` which will be called pre and post batch deletion respectively. If `post_delete`
+    returns falsey then batch processing stops.
+    """
+
+    num_deleted = 0
+
+    while True:
+        pk_batch = list(qs.values_list(pk, flat=True)[:batch_size])
+        if not pk_batch:
+            break
+
+        if pre_delete:
+            pre_delete(pk_batch)
+
+        qs.model.objects.filter(**{f"{pk}__in": pk_batch}).delete()
+        num_deleted += len(pk_batch)
+
+        if post_delete and not post_delete():
+            break
+
+    return num_deleted
+
+
 class LegacyUUIDMixin(SmartModel):
     """
     Model mixin for things with an old-style VARCHAR(36) UUID